refactor: page 로직 분리 (FE-27)
