+++ src/components/auth/LoginForm.tsx
@@ -1,0 +1,50 @@
+import { useState } from "react";
+import { LoginRequest } from "@coworkers-types";
+import { useRouter } from "next/router";
+import { useAuthStore } from "@store/useAuthStore";
+import { setAuth } from "@utils/auth";
+import { loginUser } from "../../pages/api/authApi";
+
+export default function LoginForm() {
+  const [values, setValues] = useState<LoginRequest>({
+    email: "",
+    password: "",
+  });
+  const router = useRouter();
+  const { setUser } = useAuthStore();
+
+  function handleChange(event: React.ChangeEvent<HTMLInputElement>) {
+    const { name, value } = event.target;
+
+    setValues((prevValues) => ({
+      ...prevValues,
+      [name]: value,
+    }));
+  }
+
+  async function handleSubmit(event: React.FormEvent) {
+    event.preventDefault();
+
+    const data = await loginUser(values);
+    setAuth(data);
+    setUser(data.user);
+    router.push("/team-list");
+  }
+
+  return (
+    <form onSubmit={handleSubmit}>
+      <label htmlFor="email">이메일</label>
+      <input id="email" name="email" type="text" value={values.email} onChange={handleChange} />
+
+      <label htmlFor="password">비밀번호</label>
+      <input
+        id="password"
+        name="password"
+        type="password"
+        value={values.password}
+        onChange={handleChange}
+      />
+      <button type="submit">로그인</button>
+    </form>
+  );
+}
